requirements.txt
pyserial>=3.5
prompt-toolkit>=3.0.0
